已可以和logistic系统进行交互,后续将进行优化，如秒杀、增加功能等

+++ mall-admin/src/main/java/com/admin/service/impl/ManagerServiceImpl.java
@@ -1,5 +1,6 @@
 package com.admin.service.impl;
 
+import cn.dev33.satoken.stp.StpUtil;
 import com.admin.dto.PortalGoodsDto;
 import com.admin.dto.PortalBrandDto;
 import com.admin.dto.PortalGoodsTypeDto;
@@ -17,51 +18,71 @@ public class ManagerServiceImpl implements ManagerService {
 
     @Override
     public List<PortalGoodsDto> getPortalGoods(int pageNum, int pageSize, short category) {
+        StpUtil.checkLogin();
+        StpUtil.checkPermission("manager");
         return portalServiceClient.getRegularPortalGoods(pageNum,pageSize,category);
     }
 
     @Override
     public PortalGoodsDto getPortalGoodsById(String goodsId) {
+        StpUtil.checkLogin();
+        StpUtil.checkPermission("manager");
         return  portalServiceClient.getPortalGoodsById(goodsId);
     }
 
     @Override
     public List<PortalGoodsTypeDto> getGoodsType(int pageNum, int pageSize) {
+        StpUtil.checkLogin();
+        StpUtil.checkPermission("manager");
         return portalServiceClient.getGoodsType(pageNum, pageSize);
     }
 
     @Override
     public void addGoodsType(PortalGoodsTypeDto portalGoodsTypeDto) {
+        StpUtil.checkLogin();
+        StpUtil.checkPermission("manager");
         portalServiceClient.addGoodsType(portalGoodsTypeDto);
     }
 
     @Override
     public void updateGoodsType(PortalGoodsTypeDto portalGoodsTypeDto) {
+        StpUtil.checkLogin();
+        StpUtil.checkPermission("manager");
         portalServiceClient.updateGoodsType(portalGoodsTypeDto);
     }
 
     @Override
     public void deleteGoodsType(Long typeId) {
+        StpUtil.checkLogin();
+        StpUtil.checkPermission("manager");
         portalServiceClient.deleteGoodsType(typeId);
     }
 
     @Override
     public List<PortalBrandDto> getBrand(int pageNum, int pageSize) {
+        StpUtil.checkLogin();
+        StpUtil.checkPermission("manager");
         return portalServiceClient.getBrand(pageNum, pageSize);
     }
 
     @Override
     public void addBrand(PortalBrandDto portalBrandDto) {
+        StpUtil.checkLogin();
+        StpUtil.checkPermission("manager");
         portalServiceClient.addBrand(portalBrandDto);
     }
 
     @Override
     public void updateBrand(PortalBrandDto portalBrandDto) {
+        StpUtil.checkLogin();
+        StpUtil.checkPermission("manager");
         portalServiceClient.updateBrand(portalBrandDto);
     }
 
     @Override
     public void deleteBrand(Long id) {
+        StpUtil.checkLogin();
+        StpUtil.checkPermission("manager");
         portalServiceClient.deleteBrand(id);
     }
 }
